requirements.txt
requests>=2.28
pytest>=7.0
python-dotenv>=1.0
fastapi>=0.95
uvicorn[standard]>=0.22
